Rename RemoteState to RemoteLedger and export its constructor
