+++ services/requester/requester.go
@@ -33,6 +33,7 @@ import (
 	"github.com/onflow/flow-evm-gateway/metrics"
 	"github.com/onflow/flow-evm-gateway/models"
 	errs "github.com/onflow/flow-evm-gateway/models/errors"
+	gwEvm "github.com/onflow/flow-evm-gateway/services/evm"
 	"github.com/onflow/flow-evm-gateway/storage"
 )
 
@@ -452,7 +453,7 @@ func (e *EVM) stateAt(evmHeight int64) (*state.StateDB, error) {
 	if !ok {
 		return nil, fmt.Errorf("could not convert to execution client")
 	}
-	ledger, err := newRemoteLedger(exeClient.ExecutionDataRPCClient(), cadenceHeight)
+	ledger, err := gwEvm.NewRemoteLedger(exeClient.ExecutionDataRPCClient(), cadenceHeight)
 	if err != nil {
 		return nil, fmt.Errorf("could not create remote ledger for height: %d, with: %w", cadenceHeight, err)
 	}
